Put tournament raw field in own collection (#126)
* move raw property of documents to their own raw file collection, only store pointers in documents

* add migration script to remove raw files and remove them from devdb

* get db name from config in migration script

* fix remove_raw migration script

* convert raw data to string before passing to raw file
