+++ dev/data/pending_tournaments.metadata.json
@@ -1,1 +1,1 @@
-{ "indexes" : [ { "v" : 1, "key" : { "_id" : 1 }, "name" : "_id_", "ns" : "garpr.pending_tournaments" } ] }
+{"options":{},"indexes":[{"v":1,"key":{"_id":1},"name":"_id_","ns":"garpr.pending_tournaments"}]}
