+++ dev/data/rankings.metadata.json
@@ -1,1 +1,1 @@
-{ "indexes" : [ { "v" : 1, "key" : { "_id" : 1 }, "name" : "_id_", "ns" : "garpr.rankings" }, { "v" : 1, "key" : { "time" : 1 }, "name" : "time_1", "ns" : "garpr.rankings" } ] }
+{"options":{},"indexes":[{"v":1,"key":{"_id":1},"name":"_id_","ns":"garpr.rankings"},{"v":1,"key":{"time":1},"name":"time_1","ns":"garpr.rankings"}]}
