+++ dev/data/raw_files.metadata.json
@@ -1,0 +1,1 @@
+{"options":{},"indexes":[{"v":1,"key":{"_id":1},"name":"_id_","ns":"garpr.raw_files"}]}
