+++ dev/data/tournaments.metadata.json
@@ -1,1 +1,1 @@
-{ "indexes" : [ { "v" : 1, "key" : { "_id" : 1 }, "name" : "_id_", "ns" : "garpr.tournaments" }, { "v" : 1, "key" : { "date" : 1 }, "name" : "date_1", "ns" : "garpr.tournaments" } ] }
+{"options":{},"indexes":[{"v":1,"key":{"_id":1},"name":"_id_","ns":"garpr.tournaments"},{"v":1,"key":{"date":1},"name":"date_1","ns":"garpr.tournaments"}]}
