+++ model.py
@@ -106,7 +106,7 @@ class Tournament(orm.Document):
               ('date', orm.DateTimeField()),
               ('regions', orm.ListField(orm.StringField())),
               ('url', orm.StringField()),
-              ('raw', orm.StringField()),
+              ('raw_id', orm.ObjectIDField()),
               ('matches', orm.ListField(orm.DocumentField(Match))),
               ('players', orm.ListField(orm.ObjectIDField())),
               ('orig_ids', orm.ListField(orm.ObjectIDField()))]
@@ -169,7 +169,7 @@ class Tournament(orm.Document):
             date=pending_tournament.date,
             regions=pending_tournament.regions,
             url=pending_tournament.url,
-            raw=pending_tournament.raw,
+            raw_id=pending_tournament.raw_id,
             matches=matches,
             players=players,
             orig_ids=players)
@@ -183,7 +183,7 @@ class PendingTournament(orm.Document):
               ('date', orm.DateTimeField()),
               ('regions', orm.ListField(orm.StringField())),
               ('url', orm.StringField()),
-              ('raw', orm.StringField()),
+              ('raw_id', orm.ObjectIDField()),
               ('matches', orm.ListField(orm.DocumentField(AliasMatch))),
               ('players', orm.ListField(orm.StringField())),
               ('alias_to_id_map', orm.ListField(orm.DocumentField(AliasMapping)))]
@@ -216,18 +216,27 @@ class PendingTournament(orm.Document):
 
     @classmethod
     def from_scraper(cls, type, scraper, region_id):
-        regions = [region_id]
-        return cls(
+        raw_file = RawFile(id=ObjectId(),
+                           data=str(scraper.get_raw()))
+        pending_tournament = cls(
             id=ObjectId(),
             name=scraper.get_name(),
             type=type,
             date=scraper.get_date(),
-            regions=regions,
+            regions=[region_id],
             url=scraper.get_url(),
-            raw=scraper.get_raw(),
+            raw_id=raw_file.id,
             players=scraper.get_players(),
             matches=scraper.get_matches())
+        return pending_tournament, raw_file
 
+# used to store large blobs of data (e.g. raw tournament data) so we don't
+# need to carry around tournament data as much. (might eventually be replaced
+# with something like S3)
+class RawFile(orm.Document):
+    fields = [('id', orm.ObjectIDField(required=True, load_from=MONGO_ID_SELECTOR,
+                                       dump_to=MONGO_ID_SELECTOR)),
+              ('data', orm.StringField())]
 
 class Ranking(orm.Document):
     fields = [('id', orm.ObjectIDField(required=True, load_from=MONGO_ID_SELECTOR,
